+++ src/components/shared/HorizontalGroup/index.ts
@@ -1,0 +1,1 @@
+export {HorizontalGroup} from './HorizontalGroup';